Partially introduce `query' operation; refactor old arg parsing.
Remove much of the old argument parsing, prepare the old query parameter
parsing for the new operation.
